Basic password support.
cmd/passwd lets you add a password for a username, which should
be an email given the current frontend. `NewSerializedPassword` hashes
it and `AddPassword` updates the database.

`DatabaseMiddleware` puts a `*sql.DB` into a request's context,
closing it after the request finishes.

The login handler uses `GetSerializedPassword` to retrieve a user's
password, if any, and checks it against the provided one with
`CheckPassword`.